server/requirements.txt
flask
flask-cors
google-genai
python-dotenv
# PDF & OCR
pdfplumber
pytesseract
Pillow
PyMuPDF